feat(plans): does not allow removing plans that have details

+++ src/Domains/Plans/Actions/DeletePlanByUrlAction.php
@@ -3,6 +3,7 @@
 namespace Domains\Plans\Actions;
 
 use Domains\Plans\Contracts\PlanRepository;
+use Domains\Plans\Exceptions\CannotDeletePlanWithDetailsException;
 
 class DeletePlanByUrlAction
 {
@@ -12,6 +13,12 @@ class DeletePlanByUrlAction
 
     public function __invoke(string $url): bool
     {
+        $planDetails = $this->planRepository->totalPlanDetailsByUrl($url);
+
+        if ($planDetails) {
+            throw new CannotDeletePlanWithDetailsException();
+        }
+
         return $this->planRepository->deleteByUrl($url);
     }
 }
